code/middle/reverse_linked_list_2/main_test.go
package reverse_linked_list_2

import (
	"github.com/stretchr/testify/require"
	"testing"
)

func TestName(t *testing.T) {
	t.Run("test 1", func(t *testing.T) {
		input := &ListNode{
			Val: 1,
			Next: &ListNode{
				Val: 2,
				Next: &ListNode{
					Val: 3,
					Next: &ListNode{
						Val: 4,
						Next: &ListNode{
							Val:  5,
							Next: nil,
						},
					},
				},
			},
		}
		left := 2
		right := 4
		expected := &ListNode{
			Val: 1,
			Next: &ListNode{
				Val: 4,
				Next: &ListNode{
					Val: 3,
					Next: &ListNode{
						Val: 2,
						Next: &ListNode{
							Val:  5,
							Next: nil,
						},
					},
				},
			},
		}
		actual := reverseBetween(input, left, right)
		require.Equal(t, expected, actual)
	})
}
